requirements.txt
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==2.2.5
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langdetect==1.0.9
MarkupSafe==3.0.2
matplotlib==3.7.1
moviepy==1.0.3
multidict==6.1.0
nltk==3.9.1
numpy==1.26.4
openai==0.27.7
packaging==24.2
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
proglog==0.1.10
propcache==0.2.1
pydantic==1.10.19
pydub==0.25.1
PyJWT==2.7.0
pypandoc==1.11
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.15
starlette==0.42.0
stripe==5.4.0
tenacity==9.0.0
textblob==0.17.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
